More refactoring. Created my own package
