+++ Hestia/app/src/main/java/hestia/backend/refactoring/GetRequest.java
@@ -1,0 +1,4 @@
+package hestia.backend.refactoring;
+
+public class GetRequest {
+}
